snapshots: fill up the documentation
